feat: write working Tetris mvp
Thanks to our backend implementation, getting a mvp out for a JavaFX
based frontend was relatively simple (I of course referred heavily to
the MoveSquareJavaFX prototype project I first wrote).

The hardest thing to get right at first was getting the moving down to
somehow occur in tandem with the user's keypresses, which wasn't as
simple as it looked at first. The solution was to use a separate
animation loop for moving down. Having the motions implemented in the
backend made this relatively trivial to implement.