feat: remove hard drop
This was causing a lot of problems, since our moveDown action is tied
to its own animation, not to a keypress. However, hardDrop contradicts
this notion, in part because of how it's implemented: as an
exaggerated number of moveDown actions. The piece is then frozen, but,
upon trying to move it down again, there is no collision - and so the
"freefall" state is returned! This is what we're trying to fix now.