fix: render sidebar properly before any moveDown action occurs
Before, the score and such would only render after the first moveDown
took place, understandably because the render code was only getting
called inside the relevant callback. Now I call it in the
DownwardVelocity constructor as well, so it shows up as soon as that
object is created inside the Main class somewhere.