refactor: apply 'frameRate' constant to DownwardVelocity class
Previously, the animation loop associated with this class had a rate
of one frame per second. Here, we've updated the frame rate to the one
used by the application as a whole (thirty frames per second.)
Accordingly, we've had to adjust the rate settings, to maintain the
same drop speeds as before.