fix: add keypress loop to DownwardVelocity for responsiveness
This is like the original animation loop for the same purpose in the
Main class, except this one is localized inside the
DownwardVelocity class.

Its purpose is to decouple the responsiveness of the player's spacebar
presses, from the artificially reduced rate of the downward animation
itself. In other words, pieces should start falling slowly, but this
slowness shouldn't be inherited by the keyboard interaction!