feat: add a "pause game" feature bound to 'p'
For this, we needed a new animation driver that is able to maintain a
list of all the normal gameplay drivers, such that it can pause and
unpause their individual animations.

Right now, all the keypresses are queued even while the game is
paused. Our next step is to ignore such keypresses.