refactor: use signal-based logic for user-driven events (wip)
Instead of queuing keypresses, the application now translates
user-driven events (in this case, keypresses) into events, and queues
those. These events are now what are listened for by the various
animation loops (game components).

The AnimationDriver interface has been worked into an abstract class,
in order to factor out various common fields and methods used by
animations.

What remains is to fix the quitting functionality, which is broken as
of this commit (because, while it pauses the game, it doesn't perform
its expected UI update).