refactor?: reimplement GameSignal as a sealed interface
I wanted a way of keeping track of the number of pauses available for
the animations to collectively consume, and this seemed like a good
excuse to try out Java 21's pattern matching facilities.

The stolen-pause-signal problem looks like it's been solved, but what
remains is to "debounce" the ordinary key-manipulations that involve
rotation and translation (i.e., non-move-down related).