feat: assign task of filtering null game signals to AnimationDriver
Our goal is to design a signal-based game frontend that uses a given
game state backend. Hence we want clients of AnimationDriver to
somehow be as simple as possible, with AnimationDriver itself managing
as much boilerplate and state as it can.

For example, a client animation shouldn't have to "remember" to guard
against a null signal, if AnimationDriver can do it for that
animation. (Or rather, if it can be done behind the scenes in the
persona of the abstract parent class, which I guess amounts to the
same thing.)